Extracted the People data into a different file
