+++ components/People/PeopleInfo.js
@@ -1,0 +1,79 @@
+export const PeopleData = [
+    {
+      id: 1,
+      image: "/assets/TEST/people.png",
+      name: "Belle See",
+      verified: true,
+      roles: "Founder, CommandTech"
+    },
+    {
+      id: 2,
+      image: "/assets/TEST/people.png",
+      name: "Agnes M. Reilly",
+      verified: false,
+      roles: "Founder, CommandTech"
+    },
+    {
+      id: 3,
+      image: "/assets/TEST/people.png",
+      name: "Timothy D. Jones",
+      verified: false,
+      roles: "Founder, CommandTech"
+    },
+    {
+      id: 4,
+      image: "/assets/TEST/people.png",
+      name: "Megan A. Lear",
+      verified: false,
+      roles: "Founder, CommandTech"
+    },
+    {
+      id: 5,
+      image: "/assets/TEST/people.png",
+      name: "Lillian T. Bell",
+      verified: true,
+      roles: "Founder, CommandTech"
+    },
+    {
+      id: 6,
+      image: "/assets/TEST/people.png",
+      name: "Constance A. Klemm",
+      verified: true,
+      roles: "Founder, CommandTech"
+    },
+    {
+      id: 7,
+      image: "/assets/TEST/people.png",
+      name: "Karen J. Waugh",
+      verified: true,
+      roles: "Founder, CommandTech"
+    },
+    {
+      id: 8,
+      image: "/assets/TEST/people.png",
+      name: "Lucas A. Garrison",
+      verified: true,
+      roles: "Founder, CommandTech"
+    },
+    {
+      id: 9,
+      image: "/assets/TEST/people.png",
+      name: "Harry B. Carpenter",
+      verified: true,
+      roles: "Founder, CommandTech"
+    },
+    {
+      id: 10,
+      image: "/assets/TEST/people.png",
+      name: "Harmony J. Ware",
+      verified: false,
+      roles: "Founder, CommandTech"
+    },
+    {
+      id: 11,
+      image: "/assets/TEST/people.png",
+      name: "Ruth D. Geter",
+      verified: false,
+      roles: "Founder, CommandTech"
+    }
+  ]
